refactor(infra): split session cost usage accounting
