+++ config/max-lines-baseline.txt
@@ -887,7 +887,6 @@ src/infra/restart-stale-pids.test.ts
 src/infra/restart.ts
 src/infra/run-node.test.ts
 src/infra/session-cost-usage.test.ts
-src/infra/session-cost-usage.ts
 src/infra/state-migrations.doctor.ts
 src/infra/state-migrations.runtime-state.ts
 src/infra/state-migrations.session-store.ts
